requirements.txt
numpy==2.2.5
opencv_python==4.11.0.86
Pillow==11.2.1
PyQt5==5.15.11
PyQt5_sip==12.17.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
